Update solidus buildpack to work with rails 5.2
Latest rails buildpack is going to try to understand if yarn is installed
by reading the Gemfile.lock file of the application.

Gemfile.lock is supposed to be there for all apps but we don't have one
at that stage since it's not commited and bundle install is still not
executed.

I fixed removing the yarn check, since for now we don't have it.